create health bit abstraction
